Fix compiler warnings
Added the '-Wextra' compiler flag and fix all warnings. These were
* Not checking the return value of fread().
* Some unused function parameters.
* Fall-through in switch statement.

The trigger for this and the following patches is:
    https://github.com/openMSX/openMSX/issues/1472
We fixed a bug in the openMSX diskmanipulator code. That code originally
started from this project, but has since evolved considerably. But
likely the bug in diskmanipulator is also present in this MSXtar code.

Instead of just porting the bug fix from openMSX to MSXtar, it would be
better to de-duplicate the code. We'll try to do this in (many) small
steps. First bring both code bases closer together. And then hopefully
in the end merge them into a single code base (e.g. move this project to
the Contrib/ subdirectory in the openMSX repository).