cleanup prints
Replace 'std::endl' with '\n'. Rationale: 'endl' is the combination of
'\n' plus a flush. Most of the time that flush isn't needed (and when it
is it's clearer to explicitly write 'std::flush').

Write debug prints to 'std::cerr' instead of 'std::cout'. And (by
default) 'cerr' does a flush after every printed chunk (so again no
reason to use 'endl').

I replaced combinations of direct print to 'cout' immediately followed by
call to exit() with CRITICAL_ERROR.

I replaced some direct prints to 'cout' with PRT_DEBUG, when I thought
the printed info is not useful for regular users.

I removed some PRT_DEBUG when I thought they are not (or no longer)
useful.